requirements.txt
torch==2.4.1
torchvision>=0.10.0
avalanche-lib>=0.4.0
pyyaml>=5.4.0
tqdm>=4.60.0
numpy>=1.21.0